backend/requirements.txt
beautifulsoup4==4.12.2
py-near==1.1.49
Flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
